Some page navigation
"Learn more!" all lead no where because I don't know where the listing page is located, but I have made it so that the feature is ready to go as soon as it can be.
The renterpage is connected to the listings page via the home button up top: clicking on it takes the user to listings.html.